requirements.txt
fastapi
uvicorn
pinecone-client
openai
pydantic
